refactor(context): Modified to interface-oriented
