+++ packages/artalk-sidebar/src/sidebar-component.ts
@@ -1,5 +1,5 @@
 import ArtalkComponent from 'artalk/src/lib/component'
-import Context from 'artalk/src/context'
+import Context from 'artalk/types/context'
 import { SidebarCtx } from './main'
 
 export default class Component extends ArtalkComponent {
