+++ packages/artalk-sidebar/src/sidebar-view.ts
@@ -1,4 +1,4 @@
-import Context from 'artalk/src/context'
+import Context from 'artalk/types/context'
 import * as Utils from 'artalk/src/lib/utils'
 import Comment from 'artalk/src/components/comment'
 import Component from './sidebar-component'
