+++ packages/artalk-sidebar/src/sidebar-views/comments-view.ts
@@ -1,4 +1,4 @@
-import Context from 'artalk/src/context'
+import Context from 'artalk/types/context'
 import * as Utils from 'artalk/src/lib/utils'
 import ListLite from 'artalk/src/components/list-lite'
