+++ packages/artalk-sidebar/src/sidebar-views/setting-view.ts
@@ -1,5 +1,5 @@
 import Api from 'artalk/src/api'
-import Context from 'artalk/src/context'
+import Context from 'artalk/types/context'
 import * as Utils from 'artalk/src/lib/utils'
 import Comment from 'artalk/src/components/comment'
