+++ packages/artalk/src/components/editor.ts
@@ -1,7 +1,7 @@
 import '../style/editor.less'
 
 import { CommentData } from '~/types/artalk-data'
-import Context from '../context'
+import Context from '~/types/context'
 import Component from '../lib/component'
 import * as Utils from '../lib/utils'
 import * as Ui from '../lib/ui'
