+++ packages/artalk/src/components/layer.ts
@@ -1,4 +1,4 @@
-import Context from '../context'
+import Context from '~/types/context'
 import Component from '../lib/component'
 import * as Utils from '../lib/utils'
 import * as Ui from '../lib/ui'
