+++ packages/artalk/src/components/sidebar-layer.ts
@@ -1,6 +1,6 @@
 import '../style/sidebar-layer.less'
 
-import Context from '@/context'
+import Context from '~/types/context'
 import Component from '@/lib/component'
 import * as Utils from '@/lib/utils'
 import * as Ui from '@/lib/ui'
