+++ packages/artalk/src/lib/component.ts
@@ -1,5 +1,5 @@
 import ArtalkConfig from '~/types/artalk-config'
-import Context from '../context'
+import Context from '~/types/context'
 import { I18n } from '../i18n'
 
 export default class Component {
